imz3

+++ delemp.py
@@ -1,36 +1,47 @@
 #!/data/data/com.termux/files/usr/bin/env python
-import sys
 from collections import deque
 from multiprocessing import Pool
 from pathlib import Path
-from sys import exit
-from time import perf_counter
-from dh import format_size, get_nobinary, get_size, is_binary
-from fastwalk import walk_files
+import sys
+
+from dh import format_size, get_nobinary, get_size
 from termcolor import cprint
 
 
-def process_file(filepath):
-    if filepath.is_symlink():
-        return False
+def process_file(filepath) -> int:
+    non_empty_lines: list[str] = []
+    lines: list[str] = []
+    removed: int = 0
+    content: str = ""
+    line: str = ""
+    if filepath.is_symlink() or filepath.suffix == ".bak" or not get_size(filepath):
+        return 0
     try:
-        before = get_size(filepath)
         print(f"[OK] {filepath.name}", end=" ")
         with filepath.open(
             "r+",
             encoding="utf-8",
-            errors="ignore",
+            errors="replace",
         ) as f:
-            lines = (line for line in f if line.strip())
-            content = "".join(lines)
+            lines = f.readlines()
+            for line in lines:
+                if line.strip():
+                    non_empty_lines.append(line)
+                else:
+                    removed += 1
+            content = "".join(non_empty_lines)
+            if not removed:
+                cprint(f" {removed}", "green")
+                del lines, line, removed, content, non_empty_lines
+                return 0
             f.seek(0)
             f.write(content)
             f.truncate()
-        after = get_size(filepath)
-        cprint(f"{format_size(before - after)}", "cyan")
-        return before != after
+            cprint(f" {removed}", "cyan")
+        del lines, line, content, non_empty_lines
+        return removed
     except OSError:
-        return False
+        return 0
 
 
 def main():
@@ -38,24 +49,27 @@ def main():
     dir = Path.cwd()
     start_size = get_size(dir)
     args = sys.argv[1:]
-    if args:
-        files = [Path(arg) for arg in args]
-    else:
-        files = get_nobinary(dir)
+    files = [Path(arg) for arg in args] if args else get_nobinary(dir)
     if len(files) == 1:
         process_file(files[0])
-        return
+        sys.exit(0)
+    lines_removed = 0
+    results = []
     with Pool(8) as p:
         pending = deque()
         for f in files:
             pending.append(p.apply_async(process_file, (f,)))
             if len(pending) > 16:
-                pending.popleft().get()
+                results.append(pending.popleft().get())
         while pending:
-            pending.popleft().get()
-    end_size = get_size(dir)
+            results.append(pending.popleft().get())
+    for result in results:
+        if result:
+            lines_removed += result
+    cprint(f"total empty lines removed : {lines_removed}", "green")
+    after = get_size(dir)
     print("space freed: ", end=" ")
-    cprint(f"{format_size(start_size - end_size)}", "cyan")
+    cprint(f"{format_size(start_size - after)}", "green")
 
 
 if __name__ == "__main__":
